Updated to cast thread ID to an appropriately sized integer type before casting it to (unsigned long long) for logging, prevents a GCC error when pthread_t is a pointer of a different (smaller) size than unsigned long long
FossilOrigin-Name: ca00774c8f49fa5a71c2f2f9f70847fb26993f4f

+++ filed.c
@@ -490,7 +490,7 @@ static void *filed_logging_thread(void *arg_p) {
 					break;
 			}
 			fprintf(fp, " THREAD=%llu TIME=%llu\n",
-				(unsigned long long) curr->thread,
+				(unsigned long long) ((intptr_t) curr->thread),
 				(unsigned long long) now
 			);
 			fflush(fp);
